Implement experimental mutation API
Previously, structures were interacted with by using several nested
paths for both the source and destination structure. This was very
inefficient and not intuitive to use.

The new mutation interface allows users to create a "spec", populate
that spec with data from a save (thus creating a "mutation"), then
apply that mutation to another save.

Theoretically, a mutation could also be created manually with
given values, instead of pulling the values from a save.

Other changes:

* Make SaveFiles lazily load their content
* Parallelize trickset copying
* Reorganize project:
  * Move config from cascade to cascade_gui
  * Move logger init from cascade_cli to cascade_gui
* thiserror -> anyhow
* Add automated round-trip integration testing